Ajustando os códigos da aplicação

+++ src/main/java/br/com/bank/java_bank/domain/DTO/AccountResponse.java
@@ -1,5 +1,7 @@
 package br.com.bank.java_bank.domain.DTO;
 
-public record AccountResponse(Long id, String pix, long balance) {
+import java.math.BigDecimal;
+
+public record AccountResponse(Long id, String pix, BigDecimal balance) {
 
 }
