+++ src/main/java/br/com/bank/java_bank/domain/DTO/DepositRequest.java
@@ -1,5 +1,7 @@
 package br.com.bank.java_bank.domain.DTO;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
@@ -15,6 +17,6 @@ public record DepositRequest(
                 
         @Schema(description = "Valor do saque em centavos", example = "10000")
         @Positive(message = "O valor do depósito deve ser positivo")
-        long amount) {
+        BigDecimal amount) {
 
 }
