+++ src/main/java/br/com/bank/java_bank/domain/DTO/InvestmentResponse.java
@@ -1,5 +1,7 @@
 package br.com.bank.java_bank.domain.DTO;
 
-public record InvestmentResponse(Long id, String pix, long balance, long tax) {
+import java.math.BigDecimal;
+
+public record InvestmentResponse(Long id, String pix, BigDecimal balance, BigDecimal tax) {
 
 }
